Create Dominant Army
In the medieval age, there were 
3
3 kingdoms 
�
A, 
�
B, and 
�
C. The army of these kingdom had 
�
�
N 
A
​
 , 
�
�
N 
B
​
 , and 
�
�
N 
C
​
  soldiers respectively.

You are given that an army with 
�
X soldiers can defeat an army with 
�
Y soldiers only if 
�
>
�
X>Y.

An army is said to be dominant if it can defeat both the other armies combined. For example, kingdom 
�
C's army will be dominant only if 
�
�
>
�
�
+
�
�
N 
C
​
 >N 
A
​
 +N 
B
​
 .

Determine whether any of the armies is dominant or not.

Input Format
The first line contains a single integer 
�
T - the number of test cases. Then the test cases follow.
The first and only line of each test case contains three integers 
�
�
N 
A
​
 , 
�
�
N 
B
​
 , and 
�
�
N 
C
​
  - the number of warriors in the armies of kingdoms 
�
A, 
�
B, and 
�
C respectively.
Output Format
For each test case, output YES if any of the armies is dominant. Otherwise, output NO.

You may print each character of YES and NO in uppercase or lowercase (for example, yes, yEs, Yes will be considered identical).